Refactor io/buffered.rs into submodules
